+++ benchopt/callback.py
@@ -15,6 +15,8 @@ class _Callback:
         Contains objective and data names, problem dimension, etc.
     stopping_criterion : instance of StoppingCriterion
         Object to check if we need to stop a solver.
+    wandb_cb : callable or None
+        Callback to log
 
     Attributes
     ----------
@@ -44,10 +46,11 @@ class _Callback:
         The time when exiting the callback call.
     """
 
-    def __init__(self, objective, meta, stopping_criterion):
+    def __init__(self, objective, meta, stopping_criterion, wandb_cb=None):
         self.objective = objective
         self.meta = meta
         self.stopping_criterion = stopping_criterion
+        self.wandb_cb = wandb_cb
 
         # Initialize local variables
         self.info = get_sys_info()
@@ -87,6 +90,8 @@ class _Callback:
             time=self.time_iter,
             **objective_dict, **self.info
         ))
+        if self.wandb_cb is not None:
+            self.wandb_cb(self.curve[-1])
 
         # Check the stopping criterion
         should_stop_res = self.stopping_criterion.should_stop(