+++ benchopt/runner.py
@@ -9,6 +9,7 @@ from .benchmark import _check_name_lists
 from .utils.sys_info import get_sys_info
 from .utils.files import uniquify_results
 from .utils.pdb_helpers import exception_handler
+from .utils.logging.wandb import wandb_ctx
 from .utils.terminal_output import TerminalOutput
 
 ##################################
@@ -59,7 +60,7 @@ def run_one_resolution(objective, solver, meta, stop_val):
 
 
 def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
-                   force=False, output=None, pdb=False):
+                   force=False, output=None, pdb=False, wandb=False):
     """Run all repetitions of the solver for a value of stopping criterion.
 
     Parameters
@@ -80,6 +81,10 @@ def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
         for the solver anyway. Else, use the cache if available.
     pdb : bool
         It pdb is set to True, open a debugger on error.
+    wandb : bool
+        It wandb is set to True, send the results to a wandb page. This option
+        needs to have wandb installed and configured. See the wandb
+        documentation: https://wandb.ai/quickstart.
 
     Returns
     -------
@@ -90,14 +95,15 @@ def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
     """
 
     curve = []
-    with exception_handler(output, pdb=pdb) as ctx:
+    with exception_handler(output, pdb=pdb) as ctx, \
+            wandb_ctx(meta=meta, wandb=wandb) as wandb_cb:
 
         if solver._solver_strategy == "callback":
 
             # If stopping strategy is 'callback', only call once to get the
             # results up to convergence.
             callback = _Callback(
-                objective, meta, stopping_criterion
+                objective, meta, stopping_criterion, wandb_cb=wandb_cb
             )
             solver.pre_run_hook(callback)
             callback.start()
@@ -121,6 +127,8 @@ def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
                 cost = run_one_resolution_cached(stop_val=stop_val,
                                                  **call_args)
                 curve.append(cost)
+                if wandb_cb is not None:
+                    wandb_cb(cost)
 
                 # Check the stopping criterion and update rho if necessary.
                 stop, ctx.status, stop_val = stopping_criterion.should_stop(
@@ -131,7 +139,8 @@ def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
 
 
 def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
-                   max_runs, timeout, force=False, output=None, pdb=False):
+                   max_runs, timeout, force=False, output=None, pdb=False,
+                   wandb=False):
     """Run a benchmark for a given dataset, objective and solver.
 
     Parameters
@@ -158,6 +167,10 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
         Object to format string to display the progress of the solver.
     pdb : bool
         It pdb is set to True, open a debugger on error.
+    wandb : bool
+        It wandb is set to True, send the results to a wandb page. This option
+        needs to have wandb installed and configured. See the wandb
+        documentation: https://wandb.ai/quickstart.
 
     Returns
     -------
@@ -165,7 +178,7 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
         The benchmark results.
     """
     run_one_to_cvg_cached = benchmark.cache(
-        run_one_to_cvg, ignore=['force', 'output', 'pdb']
+        run_one_to_cvg, ignore=['force', 'output', 'pdb', 'wandb']
     )
 
     # Set objective an skip if necessary.
@@ -192,6 +205,7 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
         output.set(rep=rep)
         # Get meta
         meta = dict(
+            benchmark_name=benchmark.name,
             objective_name=str(objective),
             solver_name=str(solver),
             data_name=str(dataset),
@@ -209,7 +223,7 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
             benchmark=benchmark, objective=objective,
             solver=solver, meta=meta,
             stopping_criterion=stopping_criterion,
-            force=force, output=output, pdb=pdb
+            force=force, output=output, pdb=pdb, wandb=wandb
         )
         if status in ['diverged', 'error', 'interrupted']:
             run_statistics = []
@@ -238,7 +252,7 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
                   dataset_names=None, objective_filters=None, max_runs=10,
                   n_repetitions=1, timeout=100, n_jobs=1, slurm=None,
                   plot_result=True, html=True, show_progress=True, pdb=False,
-                  output="None"):
+                  wandb=False, output="None"):
     """Run full benchmark.
 
     Parameters
@@ -279,6 +293,10 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
         If show_progress is set to True, display the progress of the benchmark.
     pdb : bool
         It pdb is set to True, open a debugger on error.
+    wandb : bool
+        It wandb is set to True, send the results to a wandb page. This option
+        needs to have wandb installed and configured. See the wandb
+        documentation: https://wandb.ai/quickstart.
     output_name : str
         Filename for the parquet output. If given, the results will
         be stored at <BENCHMARK>/outputs/<filename>.parquet.
@@ -305,7 +323,7 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
     )
     common_kwargs = dict(
         benchmark=benchmark, n_repetitions=n_repetitions, max_runs=max_runs,
-        timeout=timeout, pdb=pdb
+        timeout=timeout, pdb=pdb, wandb=wandb
     )
 
     if slurm is not None: